+++ src/app/(auth)/oauth/callback/CallbackContainer.tsx
@@ -1,0 +1,26 @@
+'use client'
+
+import { useEffect } from 'react'
+import { useSearchParams, useRouter } from 'next/navigation'
+
+export default function CallbackContainer() {
+  const searchParams = useSearchParams()
+  const router = useRouter()
+
+  useEffect(() => {
+    const token = searchParams.get('token')
+    const status = searchParams.get('status')
+
+    if (!token) return
+
+    localStorage.setItem('accessToken', token)
+
+    if (status === 'new') {
+      router.replace('/latte-chat/user/onboarding/welcome')
+    } else {
+      router.replace('/latte-chat')
+    }
+  }, [searchParams, router])
+
+  return null
+}
