+++ src/app/(auth)/oauth/callback/page.tsx
@@ -1,26 +1,12 @@
 'use client'
 
-import { useEffect } from 'react'
-import { useSearchParams, useRouter } from 'next/navigation'
+import { Suspense } from 'react'
+import CallbackContainer from './CallbackContainer'
 
 export default function CallbackPage() {
-  const searchParams = useSearchParams()
-  const router = useRouter()
-
-  useEffect(() => {
-    const token = searchParams.get('token')
-    const status = searchParams.get('status')
-
-    if (!token) return
-
-    localStorage.setItem('accessToken', token)
-
-    if (status === 'new') {
-      router.replace('/latte-chat/user/onboarding/welcome')
-    } else {
-      router.replace('/latte-chat')
-    }
-  }, [searchParams, router])
-
-  return null
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <CallbackContainer />
+    </Suspense>
+  )
 }
